requirement.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
faker
